LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_click_during_capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: click event during capture

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: Click event should be sent to the first common ancestor of the pointerdown and pointerup targets.
Click on the green box with the left button of mouse.
Press down the left button on the blue box and drag to the green box and release the button.
Repeat the two steps above once again.



Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse click target during capture
